Uses only the torsocks and torsocks2 binaries to help call tor
Torify is removed because it depends on the existence of the torsocks
binary. Usewith tor and older versions of torify will use tsocks which
has been depricated because it reveals DNS information about the user.
We default instead to using gpg over socks5 because it has options to
keep this information secret.